# contracts/openapi/tab-v1.yaml
openapi: 3.0.3
info:
  title: UTab API v1
  version: 0.1.1
  description: >
    Contract-first API used by tournament operations. This is the initial
    machine-readable contract for Phase 0.
servers:
  - url: https://tab.example.com
security:
  - sessionAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/tournaments:
    post:
      tags: [Tournaments]
      summary: Create tournament
      operationId: createTournament
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeTournament"
  /api/v1/tournaments/{id}:
    patch:
      tags: [Tournaments]
      summary: Update tournament
      operationId: updateTournament
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - $ref: "#/components/parameters/TournamentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeTournament"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Required for service-account authenticated POST/PATCH/DELETE requests.
    TournamentIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Tournament ObjectId
...
  schemas:
...
    EnvelopeTournament:
      type: object
      required: [data, errors]
      properties:
        data:
          $ref: "#/components/schemas/Tournament"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
...
    Tournament:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        style:
          type: number
    TournamentCreateRequest:
      type: object
      required: [name, style]
      properties:
        name:
          type: string
          minLength: 1
        style:
          type: number
        options:
          type: object
          additionalProperties: true
        total_round_num:
          type: integer
        current_round_num:
          type: integer
        preev_weights:
          type: array
          items:
            type: number
        auth:
          type: object
          additionalProperties: true
        user_defined_data:
          type: object
          additionalProperties: true
    TournamentPatchRequest:
      type: object
      minProperties: 1
      additionalProperties: true
...
    ErrorItem:
      type: object
      required: [name, message]
      properties:
        name:
          type: string
        message:
          type: string